ADD:: setup docker for frontend
