+++ front-end/next.config.mjs
@@ -1,4 +1,7 @@
 /** @type {import('next').NextConfig} */
-const nextConfig = {};
+const nextConfig = {
+    // distDir: "build",
+    output: 'export'
+};
 
 export default nextConfig;
