+++ front-end/nginx.conf
@@ -1,0 +1,12 @@
+server {
+    listen 80;
+
+    server_name localhost;
+
+
+    location / {
+        root /usr/share/nginx/html;
+        include /etc/nginx/mime.types;
+        try_files $uri $uri/ /index.html;
+    }
+}
